Updated readme. Small fixes
