+++ src/main/java/ru/danilarassokhin/progressive/Story.java
@@ -1,6 +1,7 @@
 package ru.danilarassokhin.progressive;
 
 import ru.danilarassokhin.progressive.component.*;
+import ru.danilarassokhin.progressive.system.StorySystem;
 
 import java.util.Map;
 
@@ -9,12 +10,9 @@ import java.util.Map;
  * Represents game story
  * <br>Story must be Singleton!
  * @param <N> Story nodes type {@link StoryNode}
- * @param <C> Story characters type {@link StoryCharacter}
- * @param <L> Story locations type {@link StoryLocation}
- * @param <I> Story item type {@link StoryItem}
- * @param <Q> Story quest type {@link StoryQuest}
+ * @param <I> Story's node id type
  */
-public interface Story<ID, N extends StoryNode> {
+public interface Story<I, N extends StoryNode> {
 
     /**
      * Returns current node
@@ -40,14 +38,14 @@ public interface Story<ID, N extends StoryNode> {
      * Returns all nodes registered in story
      * @return Map of nodes as nodeId - node
      */
-    Map<ID, N> getStoryNodes();
+    Map<I, N> getStoryNodes();
 
     /**
      * Adds node to story
      * @param id Node id to add
      * @return Empty node or null if id already exists
      */
-    N addStoryNode(ID id);
+    N addStoryNode(I id);
 
     /**
      * Sets current node in Story to {@code startNode}
@@ -56,8 +54,20 @@ public interface Story<ID, N extends StoryNode> {
      */
     N begin(N startNode);
 
+    /**
+     * Adds system to story
+     * @param system System class to add
+     * @param <S> System type
+     * @return System object or null
+     */
    <S extends StorySystem> S addSystem(Class<S> system);
 
+    /**
+     * Searches for system by it's class
+     * @param systemClass System class to search
+     * @param <S> System type
+     * @return System object or null
+     */
     <S extends StorySystem> S getSystem(Class<S> systemClass);
 
     /**
@@ -65,6 +75,6 @@ public interface Story<ID, N extends StoryNode> {
      * @param id Id to search
      * @return Node, null otherwise
      */
-    N getNodeById(ID id);
+    N getNodeById(I id);
 
 }
